requirements.txt
fastapi
uvicorn[standard]
redis
rq
jinja2
python-dotenv
requests
pyjwt
boto3
websockets
